+++ src/pages/dashboard/driver/Requests.jsx
@@ -1,4 +1,4 @@
-import { RideRequests } from "../../../components/driverDashboardFolders/RideRequest"
+import { RideRequests } from "../../../components/driverDashboardFolders/requests/RideRequest"
 import { useOutletContext } from "react-router-dom";
 
 export const Requests = () => {
